chore: Simplify code (#1004)
* feat: Refer to `https://pkg.go.dev/time` for unified comment format

* feat: Add `SetLocation` and `SetLocale` methods and unit test cases

* feat: Add `SetLocation` and `SetLocale` methods and unit test cases

* feat: Update test cases sorting

* feat: Simplify code

* feat: Add some constants

* feat: Move `SetLocation` method and unit test cases

* refactor: Remove unused errors
