+++ errors/list.go
@@ -138,7 +138,4 @@ var (
 	ValidationDuplicateRule   = New("duplicate rule name: %s")
 	ValidationEmptyData       = New("data can't be empty")
 	ValidationEmptyRules      = New("rules can't be empty")
-
-	CarbonFailedScan       = New("failed to scan %v as carbon")
-	CarbonInvalidTimestamp = New("invalid timestamp %v")
 )
